server/requirements.txt
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
psycopg2==2.8.4
flask-marshmallow==0.11.0
marshmallow-sqlalchemy==0.22.2
Flask-Cors==3.0.8
flask-script